Drop support for !Send messages
I frankly don't like it, but let's face it, rust makes this hard. We could do something like:
https://users.rust-lang.org/t/how-do-you-all-make-your-dynamic-code-send-agnostic/27567
to avoid repetition, but it would always be inconsistent (an actor has a mailbox that is
either Send or !Send) and it would split the whole library, even the interface, because
interface methods that return a boxed future would have to specify whether that is Send or not.
Currently we have even added the Send bound to Message directly.

We can always add !Send later, but it will be a separate interface. Currently it did not
seem worth the hassle for a mediocre result, since I have no use case for this anyways.